refactor: common -> mixins
