+++ .gitignore
@@ -134,3 +134,6 @@ dmypy.json
 
 # Version file
 /src/vector/version.py
+
+.pyre_configuration
+.watchmanconfig
