+++ setup.cfg
@@ -54,7 +54,6 @@ per-file-ignores =
 
 [mypy]
 files = src
-pretty = True
 python_version = 3.6
 warn_unused_configs = True
 warn_unused_ignores = True
