+++ src/vector/single/lorentz/xyzt.py
@@ -9,12 +9,12 @@ from __future__ import absolute_import, division, print_function
 import json
 from typing import TYPE_CHECKING, cast
 
-import vector.common.lorentz.xyzt
+import vector.mixins.lorentz.xyzt
 
 
 class LorentzXYZTFree(
-    vector.common.lorentz.xyzt.LorentzXYZTMethodMixin,
-    vector.common.lorentz.xyzt.LorentzXYZTDunderMixin,
+    vector.mixins.lorentz.xyzt.LorentzXYZTMethodMixin,
+    vector.mixins.lorentz.xyzt.LorentzXYZTDunderMixin,
 ):
     def __init__(self, x, y, z, t):
         # type: (float, float, float, float) -> None
